pyproject.toml
[tool.poetry]
name = "CorrectAfterglowSim"
version = "0.1.0"
description = ""
authors = ["Jonathan Schaffner <jonathan.schaffner@students.fhnw.ch>"]

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.4.0"
numpy = "^1.24.2"
matplotlib = "^3.7.1"
tifffile = "^2023.3.21"
scipy = "1.9.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
